+++ app/src/main/java/com/thecyberexpert123/androidnmap/NmapToolApplication.kt
@@ -23,6 +23,7 @@ class NmapToolApplication : Application() {
         ).addMigrations(
             AppDatabase.MIGRATION_1_2,
             AppDatabase.MIGRATION_2_3,
+            AppDatabase.MIGRATION_3_4,
         ).build()
 
         val remoteSettingsStore = RemoteSettingsStore(applicationContext)